+++ tests/e2e/game/full-game.test.js
@@ -169,28 +169,6 @@ describe("B - Full game of 31 players with all roles", () => {
                 done();
             });
     });
-    it("👪 All can't elect sheriff if votes are not set (POST /games/:id/play)", done => {
-        chai.request(server)
-            .post(`/games/${game._id}/play`)
-            .set({ Authorization: `Bearer ${token}` })
-            .send({ source: "all", action: "elect-sheriff" })
-            .end((err, res) => {
-                expect(res).to.have.status(400);
-                expect(res.body.type).to.equal("VOTES_REQUIRED");
-                done();
-            });
-    });
-    it("👪 All can't elect sheriff if votes are empty (POST /games/:id/play)", done => {
-        chai.request(server)
-            .post(`/games/${game._id}/play`)
-            .set({ Authorization: `Bearer ${token}` })
-            .send({ source: "all", action: "elect-sheriff", votes: [] })
-            .end((err, res) => {
-                expect(res).to.have.status(400);
-                expect(res.body.type).to.equal("VOTES_CANT_BE_EMPTY");
-                done();
-            });
-    });
     it("👪 All can't elect sheriff if one vote has same target and source (POST /games/:id/play)", done => {
         players = game.players;
         chai.request(server)
@@ -256,23 +234,6 @@ describe("B - Full game of 31 players with all roles", () => {
             .end((err, res) => {
                 expect(res).to.have.status(400);
                 expect(res.body.type).to.equal("CANT_VOTE_MULTIPLE_TIMES");
-                done();
-            });
-    });
-    it("👪 All can't elect sheriff if there is a tie in votes (POST /games/:id/play)", done => {
-        players = game.players;
-        chai.request(server)
-            .post(`/games/${game._id}/play`)
-            .set({ Authorization: `Bearer ${token}` })
-            .send({
-                source: "all", action: "elect-sheriff", votes: [
-                    { from: players[0]._id, for: players[1]._id },
-                    { from: players[1]._id, for: players[0]._id },
-                ],
-            })
-            .end((err, res) => {
-                expect(res).to.have.status(400);
-                expect(res.body.type).to.equal("TIE_IN_VOTES");
                 done();
             });
     });