+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 ### 🌟 Enhancements
 
 * [#176](https://github.com/antoinezanardi/werewolves-assistant-api/issues/176) - Make votes skippable.
+* [#177](https://github.com/antoinezanardi/werewolves-assistant-api/issues/177) - Random sheriff election.
 * [#178](https://github.com/antoinezanardi/werewolves-assistant-api/issues/178) - `powerless` attribute for Big Bad Wolf.
 * [#179](https://github.com/antoinezanardi/werewolves-assistant-api/issues/179) - `powerless` attribute for Pied Piper.